% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fipsCounty.R
\docType{data}
\name{fipsCounty}
\alias{fipsCounty}
\title{FIPS county data}
\format{A data frame with 3235 rows and 3 columns.}
\usage{
fipsCounty
}
\description{
FIPS county lookup table.  Adapted from census.gov \url{https://www.census.gov/geo/reference/codes/cou.html}.
}
\section{Variables}{


\itemize{
 \item \code{fips}: FIPS county code
 \item \code{county}: county name
 \item \code{state}: state abbreviation
}
}
\examples{
head(fipsCounty)
}

