% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housingData-package.R
\docType{package}
\name{housingData-package}
\alias{housingData}
\alias{housingData-package}
\title{Housing Data}
\description{
U.S. Housing Data from 2008 to 2014
}
\details{
\tabular{ll}{
Package: \tab housingData\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2015-5-13\cr
License: \tab BSD3\cr
}
}
\examples{
help(package = housingData)
}
\author{
Ryan Hafen

Maintainer: Ryan Hafen <rhafen@gmail.com>
}
\keyword{package}

