% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiCounty.R
\docType{data}
\name{wikiCounty}
\alias{wikiCounty}
\title{County Wikipedia lookup table}
\format{A data frame with 3143 rows and 5 columns.}
\usage{
wikiCounty
}
\description{
Table of links to wikipedia articles on U.S. counties.  Adapted from table here: \url{http://en.wikipedia.org/wiki/List_of_United_States_counties_and_county_equivalents}.
}
\section{Variables}{


\itemize{
 \item \code{fips}: FIPS county code
 \item \code{county}: county name
 \item \code{state}: state abbreviation
 \item \code{pop2013}: 2013 population (from wikipedia table)
 \item \code{href}: web link to wikipedia page about the county
}
}
\examples{
head(wikiCounty)
}

